Theoretical-Neuroscience/project_2/requirements.txt
ipykernel
ipywidgets
numpy
matplotlib
pandas
pillow
scipy
tqdm